+++ src/components/color_picker/color_picker.tsx
@@ -17,7 +17,10 @@ import { EuiFocusTrap } from '../focus_trap';
 import { EuiFlexGroup, EuiFlexItem } from '../flex';
 // @ts-ignore
 import { EuiFieldText } from '../form/field_text';
-import { EuiFormControlLayout } from '../form/form_control_layout';
+import {
+  EuiFormControlLayout,
+  EuiFormControlLayoutProps,
+} from '../form/form_control_layout';
 import { EuiI18n } from '../i18n';
 import { EuiPopover } from '../popover';
 import {
@@ -77,6 +80,16 @@ export interface EuiColorPickerProps
    *  Array of hex strings (3 or 6 character) to use as swatch options. Defaults to EUI visualization colors
    */
   swatches?: string[];
+
+  /**
+   * Creates an input group with element(s) coming before input. It only shows when the `display` is set to `default`.
+   */
+  prepend?: EuiFormControlLayoutProps['prepend'];
+
+  /**
+   * Creates an input group with element(s) coming after input. It only shows when the `display` is set to `default`.
+   */
+  append?: EuiFormControlLayoutProps['append'];
 }
 
 function isKeyboardEvent(
@@ -102,6 +115,8 @@ export const EuiColorPicker: FunctionComponent<EuiColorPickerProps> = ({
   readOnly = false,
   swatches = VISUALIZATION_COLORS,
   popoverZIndex,
+  prepend,
+  append,
 }) => {
   const [isColorSelectorShown, setIsColorSelectorShown] = useState(false);
   const [colorAsHsv, setColorAsHsv] = useState(
@@ -124,6 +139,7 @@ export const EuiColorPicker: FunctionComponent<EuiColorPickerProps> = ({
   }, [color, lastHex]);
 
   const classes = classNames('euiColorPicker', className);
+  const popoverClass = 'euiColorPicker__popoverAnchor';
   const panelClasses = classNames('euiColorPicker__popoverPanel', {
     'euiColorPicker__popoverPanel--pickerOnly': mode === 'picker',
     'euiColorPicker__popoverPanel--customButton': button,
@@ -131,6 +147,9 @@ export const EuiColorPicker: FunctionComponent<EuiColorPickerProps> = ({
   const swatchClass = 'euiColorPicker__swatchSelect';
   const testSubjAnchor = 'colorPickerAnchor';
   const testSubjPopover = 'colorPickerPopover';
+  const inputClasses = classNames('euiColorPicker__input', {
+    'euiColorPicker__input--inGroup': prepend || append,
+  });
 
   const handleOnChange = (hex: string) => {
     setLastHex(hex);
@@ -313,7 +332,9 @@ export const EuiColorPicker: FunctionComponent<EuiColorPickerProps> = ({
         readOnly={readOnly}
         fullWidth={fullWidth}
         compressed={compressed}
-        onKeyDown={handleToggleOnKeyDown}>
+        onKeyDown={handleToggleOnKeyDown}
+        prepend={prepend}
+        append={append}>
         <div
           // Used to pass the chosen color through to form layout SVG using currentColor
           style={{ color: showColor && color ? color : undefined }}>
@@ -325,7 +346,7 @@ export const EuiColorPicker: FunctionComponent<EuiColorPickerProps> = ({
             ]}>
             {([openLabel, closeLabel]: string[]) => (
               <EuiFieldText
-                className="euiColorPicker__input"
+                className={inputClasses}
                 onClick={handleInputActivity}
                 onKeyDown={handleInputActivity}
                 value={color ? color.toUpperCase() : ''}
@@ -360,6 +381,7 @@ export const EuiColorPicker: FunctionComponent<EuiColorPickerProps> = ({
       isOpen={isColorSelectorShown}
       closePopover={handleFinalSelection}
       zIndex={popoverZIndex}
+      className={popoverClass}
       panelClassName={panelClasses}
       display={button ? 'inlineBlock' : 'block'}
       attachToAnchor={button ? false : true}